pyproject.toml
[tool.poetry]
name = "pokras"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.5.2"
dotenv = "^0.9.9"
sqlalchemy = "^2.0.42"
pillow = "^11.3.0"
alembic = "^1.16.4"
webcolors = "^24.11.1"
pydantic = "^2.11.7"


[tool.poetry.group.dev.dependencies]
websockets = "^15.0.1"
aiohttp = "^3.12.15"
memory-profiler = "^0.61.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
